fix tests to expire cache.
Some endpoint tests were passing because tests and client are using the same session which causes the endpoint to reach for the sessions cache when fetching something from the database.

This commit adds an expire all to the session used by the application.

Tried to separate into 2 different connections one for the tests and one for the app but it seems sqlite doesnt support that. either the database locks or it is unable to rollback after the tests. When moving towards an actual database this needs to be revisited